// pms-ui/src/app/faculty/drives/components/types.ts
export interface Job {
    _id: string;
    company: string;
    drive: string;
    title: string;
    desc?: string;
    loc?: string;
    requirement?: string;
    experience?: number;
    salary?: number;
    join_date: string | Date;
    last_date: string | Date;
    contact_person?: string;
    contact_email?: string;
    additional_instructions?: string;
    form_link?: string;
    eligible_students?: string[];
    applied_students?: string[];
    selected_students?: string[];
    stage_students?: string[][];
}